Evaluation package and EACL notebook now working.

+++ deep_disfluency/evaluation/eval_utils.py
@@ -593,7 +593,7 @@ def alignment_cost(r, h, subcost=1):
     3
     """
     # initialisation
-    d = numpy.zeros((len(r)+1)*(len(h)+1), dtype=numpy.uint8)
+    d = numpy.zeros((len(r)+1)*(len(h)+1), dtype=numpy.uint16)
     d = d.reshape((len(r)+1, len(h)+1))
 
     for i in range(len(r)+1):
